% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skatz.R
\name{ASATZ}
\alias{ASATZ}
\title{SNP-set association tests using GWAS summary Z-statistics}
\usage{
ASATZ(Z, R, W = NULL, rho = c((0:5/10)^2, 0.5, 1))
}
\arguments{
\item{Z}{summary Z-statistics for a set of SNPs from GWAS}

\item{R}{SNP pairwise LD matrix}

\item{W}{SNP weights. Default to equal weights}

\item{rho}{weights for burden test}
}
\value{
\describe{
  \item{p.value}{ p-values for Adaptive test, VC, and BT }
  \item{pval}{ the list of all p-values }
  \item{rho.est}{ estimated optimal \eqn{\rho} value }
}
}
\description{
Compute p-values for the SNP-set tests (VC, burden test, adaptive test) using GWAS Z-statistics.
With less than two values of rho's, it outputs only p-values for VC and BT.
}
\examples{
R = cor(matrix(rnorm(500),100,5)*sqrt(0.8)+rnorm(100)*sqrt(0.2))
Z = rnorm(5) + 0:4
ASATZ(Z,R)
}
\references{
Guo,B., Wu,B.(2017). Statistical methods to detect novel genetic variants using publicly available GWAS summary data. tech report.
}
\keyword{ASATZ}
