% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qskat.R
\name{KAT.cnull}
\alias{KAT.cnull}
\title{Fit a null linear regression model}
\usage{
KAT.cnull(Y, X)
}
\arguments{
\item{Y}{continuous outcome}

\item{X}{covariates to be adjusted, setting X=NULL with no covariate}
}
\description{
Fit a null linear model to be used for variant set association test
}
\keyword{KAT.cnull}
