% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skatl.R
\name{KAT.null}
\alias{KAT.null}
\title{Fit a null binomial logistic regression model}
\usage{
KAT.null(D, X)
}
\arguments{
\item{D}{0-1 disease outcome}

\item{X}{covariates to be adjusted, setting X=NULL with no covariate}
}
\description{
Fit a null binomial model to be used for variant set association test
}
\keyword{KAT.null}
