% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skatl.R
\name{KAT.pval}
\alias{KAT.pval}
\title{Accurate computation of the tail probability of quadfratic forms of central normal variables}
\usage{
KAT.pval(Q.all, lambda, acc = 1e-09, lim = 1e+08)
}
\arguments{
\item{Q.all}{quadratic statistics}

\item{lambda}{coefficients of mixture of 1-DF chi-square distributions}

\item{acc}{accuracy for the Davies method}

\item{lim}{number of integration terms for the Davies method}
}
\value{
tail probabilities of quadratic forms
}
\description{
Compute the significance p-values of SKAT statistics accurately
}
\references{
Davies R.B. (1980) Algorithm AS 155: The Distribution of a Linear Combination of chi-2 Random Variables. Journal of the Royal Statistical Society. Series C (Applied Statistics), 29(3), 323-333.

P. Duchesne, P. Lafaye de Micheaux (2010) Computing the distribution of quadratic forms: Further comparisons between the Liu-Tang-Zhang approximation and exact methods. Computational Statistics and Data Analysis, 54, 858-862.

Wu,B., Guan,W., and Pankow,J.S. (2016) On efficient and accurate calculation of significance p-values for sequence kernel association test of variant set. Annals of Human Genetics, 80(2), 123-135.
}
