% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SATS.R
\name{SATS}
\alias{SATS}
\title{SNP-set association tests using GWAS Summary data}
\usage{
SATS(Z, R, W = NULL, rho = c((0:5/10)^2, 0.5, 1))
}
\arguments{
\item{Z}{summary Z-statistics for a set of SNPs from GWAS}

\item{R}{SNP pairwise LD matrix}

\item{W}{SNP weights. Default to equal weights}

\item{rho}{weights for burden test}
}
\value{
\describe{
  \item{p.value}{ p-values for AT, VC, and ST }
  \item{pval}{ the list of all p-values }
  \item{rho.est}{ estimated optimal \eqn{\rho} value }
}
}
\description{
Compute p-values for the SNP-set tests using GWAS Z-statistics: variance components test (VC), sum test (ST), and adaptive test (AT).
With less than two values of \eqn{\rho}'s, it outputs only p-values for VC and ST.
}
\examples{
R = cor(matrix(rnorm(500),100,5)*sqrt(0.8)+rnorm(100)*sqrt(0.2))
Z = rnorm(5) + 0:4
SATS(Z,R)
}
\references{
Guo,B., Wu,B.(2017). Statistical methods to detect novel genetic variants using publicly available GWAS summary data. tech report.
}
\keyword{SATS}
