% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skatl.R
\name{SKATOL}
\alias{SKATOL}
\title{Optimal sequence kernel association test (SKAT-O) for binary trait based on marginal LRT}
\usage{
SKATOL(obj, G, W.beta, rho = c(0, 0.1^2, 0.2^2, 0.3^2, 0.4^2, 0.5^2, 0.5, 1))
}
\arguments{
\item{obj}{a fitted null binomial model using KAT.null()}

\item{G}{genotype matrix, sample in rows, variant in columns}

\item{W.beta}{Beta parameters for variant weights}

\item{rho}{weights for burden test}
}
\value{
SKATOL p-value
}
\description{
Compute the significance p-value for SKAT-O based on LRT. The computational algorithm
is based on a new approach described in detail at Wu et. al (2015).
}
\examples{
library(CompQuadForm)
D = rbinom(5000,1,0.5); X = matrix(rnorm(10000),5000,2)
G = matrix(rbinom(100000,2,0.01), 5000,10)
SKATL(KAT.null(D,X), G, c(1.5,25.5))
SKATOL(KAT.null(D,X), G, c(1.5,25.5))
## library(SKAT)
## SKAT(G, SKAT_Null_Model(D~X, out_type='D'), method='davies')$p.value
## SKAT(G, SKAT_Null_Model(D~X, out_type='D'), method='optimal.adj')$p.value
}
\references{
Lee, S., Wu, M. C., and Lin, X. (2012) Optimal tests for rare variant effects in sequencing association studies. Biostatistics, 13, 762-775.

Wu, M.C., Lee, S., Cai, T., Li, Y., Boehnke, M., and Lin, X. (2011) Rare Variant Association Testing for Sequencing Data Using the Sequence Kernel Association Test (SKAT). American Journal of Human Genetics, 89, 82-93.

Wu,B., Pankow,J.S., Guan,W. (2015) Sequence kernel association analysis of rare variant set based on the marginal regression model for binary traits. Genetic Epidemiology, 39(6), 399-405.

Wu,B., Guan,W., Pankow,J.S. (2016) On efficient and accurate calculation of significance p-values for sequence kernel association test of variant set. Annals of human genetics, 80(2), 123-135.
}
\keyword{SKATO-L}
