% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qskat.R
\name{SKATOT}
\alias{SKATOT}
\title{Optimal sequence kernel association test (SKAT-O) for quantitative trait based on marginal t-statistics}
\usage{
SKATOT(obj, G, W.beta = c(1.5, 25.5), rho = c(0, 0.1^2, 0.2^2, 0.3^2, 0.4^2,
  0.5^2, 0.5, 1))
}
\arguments{
\item{obj}{a fitted null linear model using KAT.cnull()}

\item{G}{genotype matrix, sample in rows, variant in columns}

\item{W.beta}{Beta parameters for variant weights}

\item{rho}{weights for burden test}
}
\value{
SKATOT p-value
}
\description{
Compute the significance p-value for SKAT-O based on marginal t-statistics. The computational algorithm
is described in detail at Wu et. al (2015).
}
\examples{
library(CompQuadForm)
Y = rnorm(5000); X = matrix(rnorm(10000),5000,2)
G = matrix(rbinom(100000,2,0.01), 5000,10)
Y = Y + G[,1]*0.4 + G[,10]*0.3 + G[,2]*0.2
SKATT(KAT.cnull(Y,X), G, c(1.5,25.5))
SKATOT(KAT.cnull(Y,X), G, c(1.5,25.5))
## library(SKAT)
## SKAT(G, SKAT_Null_Model(Y~X, out_type='C'), method='davies')$p.value
## SKAT(G, SKAT_Null_Model(Y~X, out_type='C'), method='optimal.adj')$p.value
}
\references{
Lee, S., Wu, M. C., and Lin, X. (2012) Optimal tests for rare variant effects in sequencing association studies. Biostatistics, 13, 762-775.

Wu,B., Pankow,J.S., Guan,W. (2015) Sequence kernel association analysis of rare variant set based on the marginal regression model for binary traits. Genetic Epidemiology, 39(6), 399-405.

Wu,B., Guan,W., Pankow,J.S. (2016) On efficient and accurate calculation of significance p-values for sequence kernel association test of variant set. Annals of human genetics, in press.
}
\keyword{SKATO-T}
