% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qskat.R
\name{SKATT}
\alias{SKATT}
\title{Sequence kernel association test (SKAT) for quantitative trait based on marginal Wald-statistics}
\usage{
SKATT(obj, G, W.beta = c(1.25, 25.5))
}
\arguments{
\item{obj}{a fitted null binomial model using KAT.cnull()}

\item{G}{genotype matrix, sample in rows, variant in columns}

\item{W.beta}{Beta parameters for variant weights}
}
\value{
SKAT p-value
}
\description{
Compute the significance p-value for SKAT based on marginal Wald-statistics
}
\references{
Wu, M. C., Lee, S., Cai, T., Li, Y., Boehnke, M., and Lin, X. (2011) Rare Variant Association Testing for Sequencing Data Using the Sequence Kernel Association Test (SKAT). American Journal of Human Genetics, 89, 82-93.

Wu, M. C., Kraft, P., Epstein, M. P.,Taylor, D., M., Chanock, S. J., Hunter, D., J., and Lin, X. (2010) Powerful SNP Set Analysis for Case-Control Genome-wide Association Studies. American Journal of Human Genetics, 86, 929-942.

Duchesne, P. and Lafaye De Micheaux, P. (2010) Computing the distribution of quadratic forms: Further comparisons between the Liu-Tang-Zhang approximation and exact methods, Computational Statistics and Data Analysis, 54, 858-862.

Wu,B., Pankow,J.S., Guan,W. (2015) Sequence kernel association analysis of rare variant set based on the marginal regression model for binary traits. Genetic Epidemiology, 39(6), 399-405.

Wu,B., Guan,W., Pankow,J.S. (2016) On efficient and accurate calculation of significance p-values for sequence kernel association test of variant set. Annals of human genetics, 80(2), 123-135.
}
\keyword{SKAT}
