% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skatr.R
\name{SKATv}
\alias{SKATv}
\title{Sequence kernel association test (SKAT) using variant test statistics}
\usage{
SKATv(obj, G, W.beta = c(1, 25))
}
\arguments{
\item{obj}{a fitted null model using KAT.null() or KAT.cnull()}

\item{G}{genotype matrix, sample in rows, variant in columns}

\item{W.beta}{Beta parameters for variant weights}
}
\value{
SKAT p-value
}
\description{
Reproduce the SKAT p-value (from the SKAT R package) using marginal variant score statistics
}
\details{
We compute the SKAT based on the variant test statistics (typically of much smaller dimension), which
leads to much more efficient computations. The same algorithm as the SKAT R package
is used to compute the tail probability of 1-DF chi-square mixtures.
}
\references{
Wu, M.C., Lee, S., Cai, T., Li, Y., Boehnke, M., and Lin, X. (2011) Rare Variant Association Testing for Sequencing Data Using the Sequence Kernel Association Test (SKAT). American Journal of Human Genetics, 89, 82-93.

Wu,B., Guan,W., and Pankow,J.S. (2016) On efficient and accurate calculation of significance p-values for sequence kernel association test of variant set. Annals of Human Genetics, 80(2), 123-135.
}
\keyword{SKAT}
