% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sats.R
\name{sats}
\alias{sats}
\title{SNP-set association tests using GWAS Summary data}
\usage{
sats(Z, R, W = NULL, rho = c((0:5/10)^2, 0.5, 1))
}
\arguments{
\item{Z}{GWAS summary Z-statistics for a set of SNPs}

\item{R}{SNP pairwise LD matrix}

\item{W}{SNP weights. Default to equal weights}

\item{rho}{weights for the ST}
}
\value{
\describe{
  \item{p.value}{ p-values for AT, S2T, and ST }
  \item{pval}{ the list of all p-values }
  \item{rho.est}{ estimated optimal \eqn{\rho} value }
}
}
\description{
Compute p-values for the SNP-set tests using GWAS Z-statistics: squared sum test (S2T), sum test (ST), and adaptive test (AT).
With less than two values of \eqn{\rho}'s, it outputs only p-values for S2T and ST.
}
\examples{
R = cor(matrix(rnorm(500),100,5)*sqrt(0.8)+rnorm(100)*sqrt(0.2))
Z = rnorm(5) + 0:4
sats(Z,R)
}
\references{
Guo,B. and Wu,B.(2018). Statistical methods to detect novel genetic variants using publicly available GWAS summary data. \emph{CBC}, to appear.
}
\keyword{sats}
